requirements.txt
pyrofork
aiohttp
motor
tgcrypto
aiofiles
requests
hachoir
cloudscraper
Unidecode
pycryptodome
pytz
httpx
bs4
